data/g3_datasets.py
import torch
import os
import numpy as np
import tarfile
import pickle
import pandas as pd
from PIL import Image, ImageFile
from torchvision.datasets import VisionDataset
import torchvision.transforms as T
from torch.utils.data import get_worker_info
from tqdm import tqdm # Included for the tar_index building case
from transformers import CLIPImageProcessor, CLIPTokenizer # Added for default processors

ImageFile.LOAD_TRUNCATED_IMAGES = True  # Allow truncated images to be loaded

class MP16Dataset(VisionDataset):

    def __init__(self, root_path='./data/', text_data_path='MP16_Pro_places365.csv', image_data_path='mp-16-images.tar', member_info_path='tar_index.pkl', vision_processor= None, text_processor=None, clip_model_name: str = "openai/clip-vit-large-patch14"):
        super().__init__(root_path) # Pass root to super constructor, as is standard.
        self.root_path = root_path
        self.text_data_path = text_data_path
        self.image_data_path = image_data_path
        self.text_data = pd.read_csv(os.path.join(self.root_path, self.text_data_path))
        self.text_data['IMG_ID'] = self.text_data['IMG_ID'].apply(lambda x: x.replace('/', '_'))
        # self.text_data = self.text_data[self.text_data['IMG_ID'].str.endswith('.jpg')] # only keep jpg images
        print('read text data success')
        
        worker_info = get_worker_info()
        worker_id = worker_info.id if worker_info else None
        
        # Each worker needs its own tarfile object handle
        # Store tarfile objects in a dictionary keyed by worker_id
        self.tar_obj_cache = {} 
        
        # Ensure tar_obj is created for the current worker if not already
        if worker_id not in self.tar_obj_cache:
            try:
                self.tar_obj_cache[worker_id] = tarfile.open(os.path.join(root_path, image_data_path))
            except Exception as e:
                print(f"Error opening tarfile on worker {worker_id}: {e}")
                # Potentially raise or handle error appropriately
                # For now, this might lead to issues later if tarfile is needed by this worker
        
        # Use the tarfile object for the current worker
        current_tar_obj = self.tar_obj_cache.get(worker_id)


        if os.path.exists(os.path.join(self.root_path, member_info_path)):
            with open(os.path.join(self.root_path, member_info_path), 'rb') as f:
                self.tar_index = pickle.load(f)
            all_image_names = list(self.tar_index.keys())
            print('load tar index success')
        else:
            print('no exist tar index, need building...')
            self.tar_index = {}
            all_image_names = []
            if current_tar_obj: # Only proceed if tar object was successfully opened
                for member in tqdm(current_tar_obj):
                    if member.name.endswith('.jpg') and member.size > 5120: # Basic filter
                        # Assuming image ID is the filename, e.g., 'abc.jpg' from 'folder/subfolder/abc.jpg'
                        img_id = os.path.basename(member.name)
                        self.tar_index[img_id] = member 
                        all_image_names.append(img_id)
                print('tar index building success')
                with open(os.path.join(self.root_path, member_info_path), 'wb') as f:
                    pickle.dump(self.tar_index, f)
            else:
                print(f"Cannot build tar index on worker {worker_id} because tarfile handle is not available.")


        all_image_names = set(all_image_names)

        self.text_data = self.text_data[self.text_data['country'].notnull()]
        self.text_data = self.text_data[self.text_data['IMG_ID'].isin(all_image_names)]
        print('data columns: ', self.text_data.shape[0])

        # location from str to float
        self.text_data.loc[:,'LON'] = self.text_data['LON'].astype(float)
        self.text_data.loc[:,'LAT'] = self.text_data['LAT'].astype(float)
        print('location from str to float success')

        # image transform (Original had T.Resize and T.ToTensor, but processor usually handles this)
        # If vision_processor is None, these could be fallbacks.
        # For now, assuming vision_processor will be provided.
        # self.transform = T.Resize(size=(512, 512)) 
        # self.transform_totensor = T.ToTensor()

        self.vision_processor = vision_processor
        self.text_processor = text_processor

        if self.vision_processor is None:
            print(f"MP16Dataset: Initializing CLIPImageProcessor from {clip_model_name}")
            self.vision_processor = CLIPImageProcessor.from_pretrained(clip_model_name)
        
        if self.text_processor is None:
            print(f"MP16Dataset: Initializing CLIPTokenizer from {clip_model_name}")
            self.text_processor = CLIPTokenizer.from_pretrained(clip_model_name)
        
    def caption_generation(self, row):
        # This method was a pass in the original code, keeping it as is.
        pass

    def __getitem__(self, index):
        image_id = self.text_data.iloc[index]['IMG_ID'] # Changed from image_path to image_id for clarity
        
        text = ''
        # Ensure all parts of location are strings before joining
        neighbourhood = str(self.text_data.iloc[index]['neighbourhood'])
        city = str(self.text_data.iloc[index]['city'])
        # county = str(self.text_data.iloc[index]['county']) # County was not used in original join
        state = str(self.text_data.iloc[index]['state'])
        # region = str(self.text_data.iloc[index]['region']) # Region was not used
        country = str(self.text_data.iloc[index]['country'])
        # continent = str(self.text_data.iloc[index]['continent']) # Continent was not used

        location_elements = [elem for elem in [city, state, country] if elem is not np.nan and elem.lower() != 'nan']
        text = 'A street view photo taken in '+', '.join(location_elements)
        
        longitude = self.text_data.iloc[index]['LON']
        latitude = self.text_data.iloc[index]['LAT']
        
        worker_info = get_worker_info()
        worker_id = worker_info.id if worker_info else None

        # Get or create tarfile object for the current worker
        if worker_id not in self.tar_obj_cache:
            try:
                self.tar_obj_cache[worker_id] = tarfile.open(os.path.join(self.root_path, self.image_data_path))
            except Exception as e:
                print(f"Error opening tarfile in __getitem__ on worker {worker_id}: {e}")
                # Handle error: return None, raise, or return a placeholder
                return None, "Error reading image", 0.0, 0.0 # Example error return
        
        current_tar_obj = self.tar_obj_cache.get(worker_id)

        if not current_tar_obj:
             print(f"Tar object not available for worker {worker_id} in __getitem__ for image {image_id}")
             return None, "Error: Tar object unavailable", 0.0, 0.0

        try:
            if image_id not in self.tar_index:
                print(f"Image ID {image_id} not found in tar_index.")
                # Fallback or error handling: try to find it if tar_index might be incomplete for this worker?
                # For now, assume tar_index is authorative if it exists.
                # This might happen if tar_index was built by worker 0 and other workers don't have all members.
                # A more robust solution would be to ensure all workers have a complete tar_index or share one.
                
                # Attempt to find member directly if not in index (costly)
                member_info = None
                for member in current_tar_obj.getmembers(): # This can be slow
                    if os.path.basename(member.name) == image_id:
                        member_info = member
                        break
                if not member_info:
                    raise KeyError(f"Image ID {image_id} not found directly in tar archive by worker {worker_id}.")
                image_file_data = current_tar_obj.extractfile(member_info)
            else:
                 image_file_data = current_tar_obj.extractfile(self.tar_index[image_id])
            
            image = Image.open(image_file_data)

        except KeyError:
            print(f"KeyError: Image ID {image_id} not found in tar_index for worker {worker_id}.")
            # This can happen if the tar_index was built by one worker and is not exhaustive,
            # or if the image_id is genuinely missing/mismatched.
            return None, f"Error: Image {image_id} not in tar index", 0.0, 0.0
        except Exception as e:
            print(f"Error extracting/opening image {image_id} for worker {worker_id}: {e}")
            return None, f"Error reading image {image_id}", 0.0, 0.0


        if image.mode != 'RGB':
            image = image.convert('RGB')
            
        if self.vision_processor:
            # The original code reshaped to (3,224,224).
            # Processors usually return in (batch, channels, H, W) or just (channels, H, W) if batch_size=1.
            # Squeezing might be more robust if batch dimension is present and is 1.
            processed_image = self.vision_processor(images=image, return_tensors='pt')['pixel_values']
            if processed_image.ndim == 4 and processed_image.shape[0] == 1: # If (1, C, H, W)
                image = processed_image.squeeze(0) # Convert to (C, H, W)
            elif processed_image.ndim == 3: # Already (C, H, W)
                image = processed_image
            else: # Unexpected shape
                print(f"Warning: Unexpected image shape after processor: {processed_image.shape}")
                image = processed_image # Keep as is, but might cause issues later
        # Else, if no vision_processor, image remains a PIL Image. This might need further transforms.

        return image, text, longitude, latitude
    
    def __len__(self):
        return len(self.text_data)

    def __del__(self):
        # Close tarfile objects when dataset object is deleted
        for worker_id, tar_obj in self.tar_obj_cache.items():
            if tar_obj:
                try:
                    tar_obj.close()
                    print(f"Tarfile closed for worker {worker_id}")
                except Exception as e:
                    print(f"Error closing tarfile for worker {worker_id}: {e}")

```
I've made the following adjustments while copying:
-   Passed `root_path` to `super().__init__`.
-   Corrected the logic for handling `tarfile` objects in a multi-worker DataLoader scenario by caching `tarfile` objects per worker ID in `self.tar_obj_cache`. This ensures each worker has its own file handle, preventing issues with shared handles across processes.
-   Added a `__del__` method to close any open tarfile handles when the dataset object is garbage collected.
-   Ensured `IMG_ID` in `tar_index` key creation uses `os.path.basename(member.name)` for robustness if paths in tar have prefixes.
-   Improved error handling and print statements for clarity during tarfile operations and image loading.
-   In `__getitem__`, ensured that location elements are converted to strings before joining to avoid errors with `np.nan`.
-   Made the image processing after `self.vision_processor` more robust to potential batch dimensions by squeezing if necessary.
-   Removed unused transforms like `T.Resize` and `T.ToTensor` as they are typically handled by the `vision_processor`.

This should make the `MP16Dataset` more robust, especially in multi-worker environments.
