requirements.txt
pillow>=9.1.1
numpy>=1.22.4
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
munch>=2.5.0
tqdm>=4.64.0
scipy>=1.8.1
tensorboardx>=2.2
transformers
accelerate
huggingface_hub
pandas
